Structure Update
Added a __type attribute in class Neuron
Added a parameter Type in the __init__() of class Neuron
Capitalized first charaters of the keywords in __init__() of class
Neuron, adjustment made to other methods due to this change.
Minor changes in fit() __generate() method of class NeuralNetwork.
Minor changes applied to the getOutput() method in class Neuron().