Minor Update
Main Changes:
Removed function minmax() in the module. Re-implemented it as a method
in class NeuralNetwork called __minmax()
Implemented __rescaleInput() method in class NeuralNetwork.
Bug fix:
Fixed the getOutput() method in class Neuron. This function may return
wrong output of a neuron in input layer when its input is specified.